+++ apibuilder4elastic/custom_flow_nodes/api-builder-plugin-axway-api-management/test/testManagerANMConfigParser.js
@@ -95,7 +95,8 @@ describe('Test API-Manager configuration variations', () => {
                     },
                     "group-a": {
                         url: "http://my.group-a-api-manager.com:8075",
-                        username: "user", password: "password"
+                        username: "user", password: "password",
+                        group: "group-a"
                     }
                 }
             }
@@ -121,11 +122,14 @@ describe('Test API-Manager configuration variations', () => {
                     },
                     "group-a": {
                         url: "http://my.group-a-api-manager.com:8075",
-                        username: "user", password: "password"
+                        username: "user", password: "password",
+                        group: "group-a"
                     },
                     "group-b###us": {
                         url: "http://my.group-b-us-api-manager.com:8075",
-                        username: "user", password: "password"
+                        username: "user", password: "password",
+                        group: "group-b",
+                        region: "us"
                     }
                 }
             }
@@ -147,11 +151,14 @@ describe('Test API-Manager configuration variations', () => {
                 "configs": {
                     "group-a": {
                         url: "http://my.group-a-api-manager.com:8075",
-                        username: "user", password: "password"
+                        username: "user", password: "password",
+                        group: "group-a"
                     },
                     "group-b###us": {
                         url: "http://my.group-b-us-api-manager.com:8075",
-                        username: "user", password: "password"
+                        username: "user", password: "password",
+                        group: "group-b",
+                        region: "us"
                     }
                 }
             }